+++ src/decoding/standard_decoding.py
