+++ src/model/lightning_model.py
