--- src/tasks/reaction_prediction/__init__.py
@@ -1,1 +1,0 @@
-from tasks.reaction_prediction.data_module import ReactionPredictionDM
